requirements.txt
numpy==1.22.3
Pillow==9.1.0
pygame==2.1.2
